+++ integration-test/src/test/java/org/pastalab/fray/test/FrayTestCase.java
@@ -15,11 +15,14 @@ import org.pastalab.fray.core.scheduler.PCTScheduler;
 import org.pastalab.fray.core.scheduler.RandomScheduler;
 import org.pastalab.fray.core.scheduler.Scheduler;
 import org.pastalab.fray.test.controllers.network.reactive.success.inputstream.PipedInputStreamReadNoDeadlockMultiThread;
+import org.pastalab.fray.test.core.success.cdl.CountDownLatchInterruptNoDeadlock;
 import org.pastalab.fray.test.core.success.threadpool.ScheduledThreadPoolWorkSteal;
 
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.*;
+import java.util.concurrent.CountDownLatch;
+
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 
@@ -86,7 +89,7 @@ public class FrayTestCase {
                 new ExecutionInfo(
                         new LambdaExecutor(() -> {
                             try {
-                                PipedInputStreamReadNoDeadlockMultiThread.main(new String[]{});
+                                CountDownLatchInterruptNoDeadlock.main(new String[]{});
                             } catch (Exception e) {
                                 throw new RuntimeException(e);
                             }